\name{survey}
\docType{data}
\alias{survey}
\title{Simulated Survey Data}
\description{
  This data set is generated as an example of a social survey dataset.
  It consists of variables for unique Person and Houeshold IDs, Tenure, Age, Sex, Health, WorkStatus, Height, and Income, followed by household- and person-level initial weights (Household_w0 and Person_w0).
}
\usage{survey_data}
\format{A data frame containing 1000 persons with 11 variables}

\keyword{datasets}

